graphql: T4574: reorganize directory structure for clarity
