+++ src/services/api/graphql/bindings.py
@@ -18,9 +18,9 @@ from . graphql.queries import query
 from . graphql.mutations import mutation
 from . graphql.directives import directives_dict
 from . graphql.errors import op_mode_error
-from . utils.schema_from_op_mode import generate_op_mode_definitions
-from . utils.schema_from_config_session import generate_config_session_definitions
-from . utils.schema_from_composite import generate_composite_definitions
+from . generate.schema_from_op_mode import generate_op_mode_definitions
+from . generate.schema_from_config_session import generate_config_session_definitions
+from . generate.schema_from_composite import generate_composite_definitions
 from ariadne import make_executable_schema, load_schema_from_path, snake_case_fallback_resolvers
 
 def generate_schema():
