+++ src/services/api/graphql/generate/composite_function.py
