+++ src/services/api/graphql/generate/config_session_function.py
