+++ src/services/api/graphql/libs/op_mode.py
